requirements.txt
# Bleeding edge Django
django

# API
djangorestframework

# Configuration
django-environ

requests

# Local
factory_boy
django-test-plus
ipython==4.2.1
ipdb
